minor, actually process overhanging bits in accuify() so kick final drain in co_snarf()
This will mask out/in the bits that were overhanging in the
puncs and pat accumulators.

However, the way it's coded at the moment makes it possible to
count some occurrences of patterns twice, once when they're in the
hangover portion of an underutilitised buffer (read() returned
less bytes than the buffer could take) and then again when the
(seemingly) unprocessed bytes are moved to the beginning of the
buffer and handed to the processor again.

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>